requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
firebase-admin
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
requests
twilio
